chore: Update Nashville MTA URLs for schedule / realtime
Nashville Metro Transit Authority (the legal entity) changed the name of
its service to WeGo Public Transit in July of 2018. This meant that the
website also changed. While the URLs currently in the database are
redirecting with `HTTP 302`, at some point that is likely to stop
working.

I have also added the `license` field to the `urls` as it is allowed in
the schema. Happy to make other adjustments as needed.