Fix swift script (#325)
* Fix: updated colums, added debug mode, other fixes

Changed the Google Sheet serving as the source of submissions by providers. Columns needed to change. Added a debug mode. Other minor fixes.

* fix: Changed Google sheet source

The source for the submissions has changed, switched from a Github secret to a 1Password service account path for this URL, added a git command to remove existing and uncommitted files before starting the script.